requirements.txt
plotly
pandas
speedtest-cli